pyproject.toml
[project]
name = "llm-brand-detector"
version = "2.1.0"
description = "High-Performance LLM Brand Mention Detection Tool - Multi-AI Provider Implementation"
authors = [
    {name = "Darren Huang", email = "darrenhhuang@gmail.com"},
]
dependencies = [
    # Web framework
    "streamlit>=1.28.0",
    
    # Data validation
    "pydantic>=2.5.0",
    
    # AI providers
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "google-generativeai>=0.3.0",
    
    # HTTP client
    "httpx>=0.25.0",
    
    # Data processing
    "pandas>=2.1.0",
    
    # Async file operations
    "aiofiles>=23.2.1",
    
    # Logging
    "loguru>=0.7.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
llm-brand-detector = "firegeo.streamlit_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/firegeo"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src/firegeo --cov-report=html --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"